feat: Move to views folder
