+++ apps/blog/.env
@@ -1,1 +1,4 @@
 NEXT_PUBLIC_APEX_URL = "https://pancakeswap.finance"
+
+NEXT_PUBLIC_STRAPI_API_URL = ""
+STRAPI_PUBLIC_API_TOKEN = ""
