+++ apps/blog/.env.example
@@ -1,0 +1,2 @@
+NEXT_PUBLIC_STRAPI_API_URL =
+STRAPI_PUBLIC_API_TOKEN =
