+++ apps/blog/pages/blog/index.tsx
@@ -1,10 +1,5 @@
-import { FC } from 'react'
 import { SWRConfig } from 'swr'
-import { Box, Flex, Button } from '@pancakeswap/uikit'
-import NewBlog from 'components/Blog/NewBlog'
-import ChefsChoice from 'components/ChefsChoice'
-import { useTranslation } from '@pancakeswap/localization'
-import NextLink from 'next/link'
+import Blog from 'views/Blog'
 
 // import { InferGetServerSidePropsType } from 'next'
 
@@ -26,24 +21,6 @@ const BlogPage = ({ fallback }: { fallback: () => void }) => {
     <SWRConfig value={{ fallback }}>
       <Blog />
     </SWRConfig>
-  )
-}
-
-const Blog: FC = () => {
-  const { t } = useTranslation()
-
-  return (
-    <Box width="100%" mb="150px">
-      <NewBlog />
-      <ChefsChoice />
-      <Flex justifyContent="center" m="50px auto">
-        <NextLink href="/blog/article" passHref>
-          <Button scale="md" variant="secondary">
-            {t('More')}
-          </Button>
-        </NextLink>
-      </Flex>
-    </Box>
   )
 }
