+++ apps/blog/views/Blog/components/Article/SingleArticle/SimilarArticles.tsx
@@ -1,9 +1,9 @@
 import { Swiper, SwiperSlide } from 'swiper/react'
 import { Button, Flex } from '@pancakeswap/uikit'
-import BlogCard from 'components/Blog/BlogCard'
+import BlogCard from 'views/Blog/components/BlogCard'
 import { useTranslation } from '@pancakeswap/localization'
 import { Autoplay, Grid } from 'swiper'
-import ArticleView from 'components/Article/ArticleView'
+import ArticleView from 'views/Blog/components/Article/ArticleView'
 import NextLink from 'next/link'
 import 'swiper/css/grid'
 import 'swiper/css/bundle'
